Implement Counter (example 1)
This implements the first example: the counter. This is such a simple
problem that I struggled a bit with structuring this. I didn't want to
play code golf here, even though most of the code could be inlined in
the `main` function.

I decided to try to stick to a more "standard" approach, even though
it's not strictly necessary for something this simple. This includes
defining a `type alias Flags = ()` alias for readability, and having
separate standalone `initialModel`, `update`, and `view` functions.

https://eugenkiss.github.io/7guis/tasks#counter