Introduce custom unit types
I'm using floats to represent multiple different types of values so I
created custom wrapper types `Celsius` and `Farenheit` to allow me (and
the compiler) to distinguish between the two.

I thought about introducing a unit of measure library like
https://package.elm-lang.org/packages/ianmackenzie/elm-units/latest but
decided against it since the implementation of unit conversions seemed
core to the exercise.

For more on types and units-of-measure, see
[A Number by any Other Name](https://www.youtube.com/watch?v=WnTw0z7rD3E )