Store string values for temperature converter
Previously we stored a single float celsius value as the source of truth
for both fields. This led to two issues:

Firstly, storing a parsed float was a lossy operation since we no longer
knew if the user typed 1.0 or 1.00. We used the float for the `value`
attribute of the input when re-rendering. This could cause user input to
be truncated as they typed. An example of this can be seen here:
https://ellie-app.com/dMYdbSsYHkxa1. To get around the problem, I
changed the model to store a string rather than a float.

Secondly, a closer reading of the spec revealed that the two inputs
aren't as as coupled as I initially thought:

> When the user enters a non-numerical string into C the value in F is
> not updated and vice versa.

This means that we can't always derive the values of both fields from a
single value. Instead, I capture two values now.

Units of measure
----------------

Floats only exist for an instant as an intermediate value as we convert
one string to another. Given that this is the only use of these numbers,
I opted not to keep the custom `Celsius` and `Farenheit` types and just
use raw floats. If the floats get used or stored anywhere else, I would
probably want to bring back these types though.